MBX-2088 Logger refactoring
