+++ Mindbox/Utilities/Logger/Logger.swift
@@ -1,99 +1,210 @@
 //
-//  Loger.swift
+//  Mindbox.swift
 //  Mindbox
 //
-//  Created by Mikhail Barilov on 13.01.2021.
+//  Created by Mikhail Barilov on 25.01.2021.
 //  Copyright © 2021 Mikhail Barilov. All rights reserved.
 //
 
 import Foundation
-import os
 
-@objcMembers
-public class MBLogger: NSObject {
+class Logger {
+    private typealias Meta = (filename: String, line: Int, funcName: String)
+    private typealias Borders = (start: String, end: String)
     
-    /**
-     ### Levels:
-     0. debug - 🪲
-     1. info - ℹ️
-     2. default - 💡
-     3. error - ‼️
-     4. fault - ⚠️
-     5. none
-     
-     - Note: `.error` by default; `.none` for disable logging
-     */
-    public var logLevel: LogLevel = .error
+    private static let subsystem: String = "cloud.Mindbox"
+    private static func log(message: String,
+                            level: LogLevel,
+                            category: LogCategory,
+                            meta: Meta,
+                            borders: Borders,
+                            subsystem: String? = nil) {
+        var header = "\n"
+        header += subsystem ?? self.subsystem + " "
+        header += category.emoji + " "
+        header += level.emoji + " "
+        header += Date().toString() + " "
+
+        header += "\n[\(sourceFileName(filePath: meta.filename))]:\(meta.line) \(meta.funcName)"
         
-    private enum ExecutionMethod {
-        case sync(lock: NSRecursiveLock)
-        case async(queue: DispatchQueue)
+        Mindbox.logger.log(
+            level: level,
+            message: borders.start + message + borders.end,
+            category: category,
+            subsystem: subsystem ?? "cloud.Mindbox"
+        )
     }
     
-    private let executionMethod: ExecutionMethod
+    static func error(_ error: MindboxError,
+                      level: LogLevel = .error,
+                      category: LogCategory = .network,
+                      fileName: String = #file,
+                      line: Int = #line,
+                      funcName: String = #function
+    ) {
+        var logMessage: String = ""
+
+        switch error {
+        case .validationError:
+            logMessage = logMessage + "\n[validationError: \(error.errorDescription ?? "No description")]"
+        case let .protocolError(e):
+            logMessage = logMessage + "\n[status: \(e.status)]"
+            logMessage = logMessage + "\n[responseError: \(error.errorDescription ?? "No description")]"
+            logMessage = logMessage + "\n[httpStatusCode: \(e.httpStatusCode)]"
+        case let .serverError(e):
+            logMessage = logMessage + "\n\(e.description)"
+        case let .internalError(e):
+            logMessage = logMessage + "\n[key: \(e.errorKey)]"
+            if let rawError = e.rawError {
+                logMessage = logMessage + "\n[message: \(rawError.localizedDescription)]"
+            }
+        case let .invalidResponse(e):
+            guard let e = e else { return }
+            logMessage = logMessage + "\n[response: \(String(describing: e))]"
+        case .connectionError:
+            logMessage = logMessage + "\n[connectionError]"
+        case let .unknown(e):
+            logMessage = logMessage + "\n[error: \(e.localizedDescription)]"
+        }
+
+        if logMessage.isEmpty { return }
+
+        let message = "LogManager: \n--- Error --- \(String(describing: logMessage)) \n--- End ---\n"
+        
+        let meta: Meta = (fileName, line, funcName)
+        let borders: Borders = ("", "\n")
+        log(message: message, level: .debug, category: .network, meta: meta, borders: borders)
+    }
     
-    public override init() {
-        #if DEBUG
-            executionMethod = .sync(lock: NSRecursiveLock())
-        #else
-            executionMethod = .async(queue: DispatchQueue(label: "Mindbox.serial.log.queue", qos: .utility))
-        #endif
-        super.init()
+    static func network(request: URLRequest,
+                        httpAdditionalHeaders: [AnyHashable: Any]? = nil,
+                        fileName: String = #file,
+                        line: Int = #line,
+                        funcName: String = #function) {
+        
+        let urlString = request.url?.absoluteString ?? ""
+        let components = NSURLComponents(string: urlString)
+
+        let method = request.httpMethod != nil ? "\(request.httpMethod!)" : ""
+        let path = "\(components?.path ?? "")"
+        let query = "\(components?.query ?? "")"
+        let host = "\(components?.host ?? "")"
+
+        var requestLog = ""
+        requestLog += "[Url]: \(urlString)"
+        requestLog += "[Method]: \(method) \(path)?\(query) HTTP/1.1\n"
+        requestLog += "[Host]: \(host)\n"
+
+        requestLog += "\n"
+        requestLog += "[Headers]: \n"
+        httpAdditionalHeaders?.forEach {
+            requestLog += "\($0.key): \($0.value)\n"
+        }
+        for (key, value) in request.allHTTPHeaderFields ?? [:] {
+            requestLog += "\(key): \(value)\n"
+        }
+
+        if let body = request.httpBody {
+            requestLog += "\n[Body]: \n"
+            let bodyString = NSString(data: body, encoding: String.Encoding.utf8.rawValue) ?? "Can't render body; not utf8 encoded"
+            requestLog += "\n\(bodyString)\n"
+        }
+        
+        let message = requestLog
+        let meta: Meta = (fileName, line, funcName)
+        let borders: Borders = ("\n[---------- OUT ---------->\n", "\n------------------------>]")
+        log(message: message, level: .debug, category: .network, meta: meta, borders: borders)
+    }
+    
+    static func response(data: Data?,
+                         response: URLResponse?,
+                         error: Error?,
+                         fileName: String = #file,
+                         line: Int = #line,
+                         funcName: String = #function) {
+        let urlString = response?.url?.absoluteString
+        let components = NSURLComponents(string: urlString ?? "")
+
+        let path = "\(components?.path ?? "")"
+        let query = "\(components?.query ?? "")"
+
+        var responseLog = ""
+        if let urlString = urlString {
+            responseLog += "[Url]: \(urlString)\n"
+        }
+
+        if let statusCode = (response as? HTTPURLResponse)?.statusCode {
+            responseLog += "[Status code]: HTTP \(statusCode) \(path)?\(query)\n"
+        }
+        if let host = components?.host {
+            responseLog += "[Host]: \(host)\n"
+        }
+
+//        for (key, value) in (response as? HTTPURLResponse)?.allHeaderFields ?? [:] {
+//            responseLog += "\(key): \(value)\n"
+//        }
+        
+        if let body = data,
+           let object = try? JSONSerialization.jsonObject(with: body, options: []),
+           let data = try? JSONSerialization.data(withJSONObject: object, options: [.prettyPrinted]),
+           let prettyPrintedString = NSString(data: data, encoding: String.Encoding.utf8.rawValue) {
+            responseLog += "[Body]: \n\(prettyPrintedString)\n"
+        }
+//
+//        if let body = data {
+//            responseLog += "[Body]: \n\(String(data: body, encoding: .utf8) ?? "")\n"
+//        }
+        
+        var level: LogLevel = .debug
+        if let error = error {
+            responseLog += "\n[Error]: \(error.localizedDescription)\n"
+            level = .error
+        }
+        let message = responseLog
+        let meta: Meta = (fileName, line, funcName)
+        let borders: Borders = ("\n[<---------- IN ----------\n", "\n<------------------------]")
+        log(message: message, level: level, category: .network, meta: meta, borders: borders)
+    }
+    
+    static func common(message: String,
+                       level: LogLevel = .debug,
+                       category: LogCategory = .general,
+                       subsystem: String? = nil,
+                       fileName: String = #file,
+                       line: Int = #line,
+                       funcName: String = #function) {
+        let meta: Meta = (fileName, line, funcName)
+        let borders: Borders = ("", "\n")
+        log(message: message, level: level, category: category, meta: meta, borders: borders, subsystem: subsystem)
+    }
+}
+
+internal extension Date {
+    func toString() -> String {
+        return Date.dateFormatter.string(from: self as Date)
     }
 
-    func log(level: LogLevel, message: String, category: LogCategory, subsystem: String) {
-        guard logLevel.rawValue <= level.rawValue else {
-            return
-        }
-        let categories: [LogCategory] = LogCategory.allCases
-        guard categories.contains(category) else {
-            return
-        }
-        let writer = makeWriter(subsystem: subsystem, category: category)
-        switch executionMethod {
-        case let .async(queue: queue):
-            queue.async { writer.writeMessage(message, logLevel: level) }
-        case let .sync(lock: lock):
-            lock.lock(); defer { lock.unlock() }
-            writer.writeMessage(message, logLevel: level)
-        }
+    func toFullString() -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZ"
+        return dateFormatter.string(from: self as Date)
     }
     
-    /**
-     Method to write log in Xcode debug output as well in Console.app.
-     
-     - Important:
-     To filter SDK logs in _Console.app_ use __subsystem__:
-     your bundle identifier which is defined by the CFBundleIdentifier key in the bundle’s information property list.
-     
-     - Warning:
-     If sdk could't find your bundleID, your logs will be under
-     subsystem __"cloud.Mindbox.UndefinedHostApplication"__
-     
-     - Parameters:
-        - level: LogLevel of the log
-        - message: Log message
-        - fileName: By default it uses #file
-        - line: By default it uses #line
-        - funcName: By default it uses #function
-     */
-    public func log(
-        level: LogLevel,
-        message: String,
-        fileName: String = #file,
-        line: Int = #line,
-        funcName: String = #function
-    ) {
-        Log(message)
-            .meta(filename: fileName, line: line, funcName: funcName)
-            .subsystem(Mindbox.shared.container?.utilitiesFetcher.hostApplicationName ?? "cloud.Mindbox.UndefinedHostApplication")
-            .level(level)
-            .category(.general)
-            .make()
+    static var dateFormatter: DateFormatter {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "hh:mm:ss.SSSS"
+        return formatter
     }
-    
-    private func makeWriter(subsystem: String, category: LogCategory) -> LogWriter {
-        return OSLogWriter(subsystem: subsystem, category: category.rawValue.capitalized)
+}
+
+// MARK: - Private Functions
+private extension Logger {
+    /// Extract the file name from the file path
+    ///
+    /// - Parameter filePath: Full file path in bundle
+    /// - Returns: File Name with extension
+    static func sourceFileName(filePath: String) -> String {
+        let components = filePath.components(separatedBy: "/")
+        return components.isEmpty ? "" : components.last!
     }
-    
 }
