+++ Mindbox/Utilities/Logger/MBLogger.swift
@@ -1,0 +1,100 @@
+//
+//  Loger.swift
+//  Mindbox
+//
+//  Created by Mikhail Barilov on 13.01.2021.
+//  Copyright © 2021 Mikhail Barilov. All rights reserved.
+//
+
+import Foundation
+import os
+
+@objcMembers
+public class MBLogger: NSObject {
+    
+    /**
+     ### Levels:
+     0. debug - 🪲
+     1. info - ℹ️
+     2. default - 💡
+     3. error - ‼️
+     4. fault - ⚠️
+     5. none
+     
+     - Note: `.error` by default; `.none` for disable logging
+     */
+    public var logLevel: LogLevel = .error
+        
+    private enum ExecutionMethod {
+        case sync(lock: NSRecursiveLock)
+        case async(queue: DispatchQueue)
+    }
+    
+    private let executionMethod: ExecutionMethod
+    
+    public override init() {
+        #if DEBUG
+            executionMethod = .sync(lock: NSRecursiveLock())
+        #else
+            executionMethod = .async(queue: DispatchQueue(label: "Mindbox.serial.log.queue", qos: .utility))
+        #endif
+        super.init()
+    }
+
+    func log(level: LogLevel, message: String, category: LogCategory, subsystem: String) {
+        guard logLevel.rawValue <= level.rawValue else {
+            return
+        }
+        let categories: [LogCategory] = LogCategory.allCases
+        guard categories.contains(category) else {
+            return
+        }
+        let writer = makeWriter(subsystem: subsystem, category: category)
+        switch executionMethod {
+        case let .async(queue: queue):
+            queue.async { writer.writeMessage(message, logLevel: level) }
+        case let .sync(lock: lock):
+            lock.lock(); defer { lock.unlock() }
+            writer.writeMessage(message, logLevel: level)
+        }
+    }
+    
+    /**
+     Method to write log in Xcode debug output as well in Console.app.
+     
+     - Important:
+     To filter SDK logs in _Console.app_ use __subsystem__:
+     your bundle identifier which is defined by the CFBundleIdentifier key in the bundle’s information property list.
+     
+     - Warning:
+     If sdk could't find your bundleID, your logs will be under
+     subsystem __"cloud.Mindbox.UndefinedHostApplication"__
+     
+     - Parameters:
+        - level: LogLevel of the log
+        - message: Log message
+        - fileName: By default it uses #file
+        - line: By default it uses #line
+        - funcName: By default it uses #function
+     */
+    public func log(
+        level: LogLevel,
+        message: String,
+        fileName: String = #file,
+        line: Int = #line,
+        funcName: String = #function
+    ) {
+        let subsystem = Mindbox.shared.container?.utilitiesFetcher.hostApplicationName ?? "cloud.Mindbox.UndefinedHostApplication"
+        Logger.common(message: message,
+                      level: level,
+                      category: .general,
+                      subsystem: subsystem,
+                      fileName: fileName,
+                      line: line,
+                      funcName: funcName)
+    }
+    
+    private func makeWriter(subsystem: String, category: LogCategory) -> LogWriter {
+        return OSLogWriter(subsystem: subsystem, category: category.rawValue.capitalized)
+    }
+}
